requirements.txt
fastapi
uvicorn
pymysql
passlib[bcrypt]
python-jose